added validations

+++ part-2/laravel-backend/app/Http/Controllers/AuthorController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Author;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class AuthorController extends Controller
 {
@@ -12,6 +13,10 @@ class AuthorController extends Controller
      */
     public function index()
     {
+        if (Author::all()->isEmpty()) {
+            return response()->json(['error' => 'No authors found'], 404);
+        }
+
         return response()->json(Author::all());
     }
 
@@ -20,9 +25,13 @@ class AuthorController extends Controller
      */
     public function store(Request $request)
     {
-        $request -> validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:50',
         ]);
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 400);
+        }
 
         $author = Author::create($request->all());
 
@@ -34,6 +43,14 @@ class AuthorController extends Controller
      */
     public function show(string $id)
     {
+        if (!is_numeric($id)) {
+            return response()->json(['error' => 'Invalid author ID'], 400);
+        }
+
+        if (!Author::find($id)) {
+            return response()->json(['error' => 'Author not found'], 404);
+        }
+
         return response()->json(Author::findOrFail($id));
     }
 
@@ -42,11 +59,19 @@ class AuthorController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $request->validate([
+        $validator = Validator::make($request->all(), [
             'name' => 'sometimes|required|string|max:50',
         ]);
 
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 400);
+        }
+
         $author = Author::findOrFail($id);
+
+        if (!$author) {
+            return response()->json(['error' => 'Author not found'], 404);
+        }
         $author->update($request->all());
 
         return response()->json($author);
@@ -57,9 +82,17 @@ class AuthorController extends Controller
      */
     public function destroy(string $id)
     {
-        $author = Author::findOrFail($id);
-        $author->delete();
+        if (!is_numeric($id)) {
+            return response()->json(['error' => 'Invalid author ID'], 400);
+        }
 
+        $author = Author::findOrFail($id);
+
+        if (!$author) {
+            return response()->json(['error' => 'Author not found'], 404);
+        }
+
+        $author->delete();
         return response()->json(['message' => 'Author deleted successfully'], 200);
 
     }
